login state holds active mentor
