+++ client/src/store/slices/login/index.ts
@@ -4,11 +4,26 @@ Permission to use, copy, modify, and distribute this software and its documentat
 
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
-import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import * as api from "api";
-import { User } from "types";
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { LoadingError } from "hooks/graphql/loading-reducer";
+import { Mentor, User } from "types";
+import {
+  googleLogin,
+  login,
+  saveMentorSubjects,
+  saveThumbnail,
+  saveMentor,
+  loadMentor,
+} from "./actions";
 
 /** Store */
+
+export interface ActiveMentor {
+  data?: Mentor;
+  mentorStatus: MentorStatus;
+  error?: LoadingError;
+  userLoadedBy?: string;
+}
 
 export enum LoginStatus {
   NONE = 0,
@@ -18,50 +33,29 @@ export enum LoginStatus {
   FAILED = 4,
 }
 
+export enum MentorStatus {
+  NONE = 0,
+  LOADING = 1,
+  SAVING = 2,
+  SUCCEEDED = 3,
+  FAILED = 4,
+}
+
 export interface LoginState {
   accessToken?: string;
   loginStatus: LoginStatus;
   user?: User;
+  activeMentor: ActiveMentor;
 }
 
 const initialState: LoginState = {
   loginStatus: LoginStatus.NONE,
+  activeMentor: {
+    mentorStatus: MentorStatus.NONE,
+  },
 };
 
 /** Actions */
-
-export const googleLogin = createAsyncThunk(
-  "login/googleLogin",
-  async (accessToken: string, { rejectWithValue }) => {
-    try {
-      const googleToken = await api.loginGoogle(accessToken);
-      return await api.login(googleToken.accessToken);
-    } catch (err) {
-      console.error(err.response.data);
-      return rejectWithValue(err.response.data);
-    }
-  }
-);
-
-export const login = createAsyncThunk(
-  "login/login",
-  async (accessToken: string, { rejectWithValue }) => {
-    try {
-      accessTokenStore(accessToken);
-      return await api.login(accessToken);
-    } catch (err) {
-      console.error(err.response.data);
-      return rejectWithValue(err.response.data);
-    }
-  }
-);
-
-export const logout =
-  () =>
-  (dispatch: (arg0: { payload: undefined; type: string }) => void): void => {
-    accessTokenClear();
-    dispatch(onLogout());
-  };
 
 /** Reducer */
 
@@ -73,6 +67,16 @@ export const loginSlice = createSlice({
       delete state.user;
       delete state.accessToken;
       state.loginStatus = LoginStatus.NOT_LOGGED_IN;
+    },
+    loadingInProgress: (state, action: PayloadAction<LoginState>) => {
+      state.activeMentor.mentorStatus = MentorStatus.LOADING;
+      state.activeMentor.userLoadedBy = action.payload.user?._id;
+    },
+    clearError: (state) => {
+      delete state.activeMentor.error;
+    },
+    updateMentor: (state, action: PayloadAction<Mentor>) => {
+      state.activeMentor.data = action.payload;
     },
   },
   extraReducers: (builder) => {
@@ -102,35 +106,69 @@ export const loginSlice = createSlice({
       .addCase(googleLogin.rejected, (state) => {
         delete state.user;
         state.loginStatus = LoginStatus.FAILED;
+      })
+      .addCase(loadMentor.fulfilled, (state, action) => {
+        if (action.payload.isCancelled || !action.payload.result) {
+          return;
+        }
+        state.activeMentor.data = action.payload.result;
+        state.activeMentor.mentorStatus = MentorStatus.SUCCEEDED;
+      })
+      .addCase(loadMentor.rejected, (state) => {
+        delete state.activeMentor.data;
+        state.activeMentor.error = {
+          message: "failed to load mentor",
+          error: loadMentor.rejected.name,
+        };
+        state.activeMentor.mentorStatus = MentorStatus.FAILED;
+      })
+      .addCase(saveMentor.pending, (state) => {
+        state.activeMentor.mentorStatus = MentorStatus.SAVING;
+      })
+      .addCase(saveMentor.fulfilled, (state, action) => {
+        state.activeMentor.data = action.payload as Mentor;
+        state.activeMentor.mentorStatus = MentorStatus.SUCCEEDED;
+      })
+      .addCase(saveMentor.rejected, (state) => {
+        state.activeMentor.mentorStatus = MentorStatus.FAILED;
+        state.activeMentor.error = {
+          message: "failed to save mentor",
+          error: saveMentor.rejected.name,
+        };
+      })
+      .addCase(saveThumbnail.pending, (state) => {
+        state.activeMentor.mentorStatus = MentorStatus.SAVING;
+      })
+      .addCase(saveThumbnail.fulfilled, (state, action) => {
+        if (state.activeMentor.data) {
+          state.activeMentor.data.thumbnail = String(action.payload);
+        }
+        state.activeMentor.mentorStatus = MentorStatus.SUCCEEDED;
+      })
+      .addCase(saveThumbnail.rejected, (state) => {
+        state.activeMentor.mentorStatus = MentorStatus.FAILED;
+        state.activeMentor.error = {
+          message: "failed to save mentor",
+          error: saveThumbnail.rejected.name,
+        };
+      })
+      .addCase(saveMentorSubjects.pending, (state) => {
+        state.activeMentor.mentorStatus = MentorStatus.SAVING;
+      })
+      .addCase(saveMentorSubjects.fulfilled, (state, action) => {
+        state.activeMentor.data = action.payload as Mentor;
+        state.activeMentor.mentorStatus = MentorStatus.SUCCEEDED;
+      })
+      .addCase(saveMentorSubjects.rejected, (state) => {
+        state.activeMentor.mentorStatus = MentorStatus.FAILED;
+        state.activeMentor.error = {
+          message: "failed to save subjects",
+          error: saveMentorSubjects.rejected.name,
+        };
       });
   },
 });
 
-const { onLogout } = loginSlice.actions;
+export const { onLogout, clearError, updateMentor } = loginSlice.actions;
 
 export default loginSlice.reducer;
-
-/** Helpers */
-
-const ACCESS_TOKEN_KEY = "accessToken";
-
-export function accessTokenGet(): string {
-  if (typeof window === "undefined") {
-    return "";
-  }
-  return localStorage.getItem(ACCESS_TOKEN_KEY) || "";
-}
-
-function accessTokenStore(accessToken: string): void {
-  if (typeof window === "undefined") {
-    return;
-  }
-  localStorage.setItem(ACCESS_TOKEN_KEY, accessToken);
-}
-
-function accessTokenClear(): void {
-  if (typeof window === "undefined") {
-    return;
-  }
-  localStorage.removeItem(ACCESS_TOKEN_KEY);
-}
