+++ client/src/store/store.ts
@@ -8,14 +8,12 @@ import { logger } from "redux-logger";
 import { configureStore } from "@reduxjs/toolkit";
 import loginReducer from "./slices/login";
 import configReducer from "./slices/config";
-import mentorReducer from "./slices/mentor";
 
 export const store = configureStore({
   middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
   reducer: {
     login: loginReducer,
     config: configReducer,
-    mentor: mentorReducer,
   },
 });
